move priceFetch function to server.js
