+++ client/src/stock-card.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import priceFetch from './price-fetch.js';
 
 let StockCard = () =>{
     return <div className="stock-card">
@@ -7,7 +6,7 @@ let StockCard = () =>{
         <div className="stock-info-short">
             <h3 className="company-name">Company Name</h3>
             <p className="company-description">Short company description.</p>
-            <p className="current-stock-price">Stock Price: {priceFetch()}</p>
+            <p className="current-stock-price">Stock Price: $0</p>
         </div>
         <div className="buy-info">
             <input className="qty-input" type="text" placeholder="Qty"/>
